seperate Trie and TrieNode class to trie.py
